QP: Put statement parsers in a separate package.
