+++ src/QueryProcessor/StatementParsers/DeleteParser.java
@@ -1,12 +1,13 @@
-package QueryProcessor;
+package QueryProcessor.StatementParsers;
 
 import Predicate.Predicate;
+import QueryProcessor.DMLStatement;
+import QueryProcessor.SQLStatement;
 import RheaDB.DBError;
 
 import java.util.Vector;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 
 public class DeleteParser extends StatementParser {
